pyproject.toml
[project]
name = "parami"
version = "0.1.0"
description = "Tools to place semantic constraints on configuration files."
readme = "README.md"
authors = [
    { name = "Danny keizer", email = "danny.keizer@protonmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "mdpython>=0.0.29",
    "python-hcl2>=7.3.1",
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
    "toml>=0.10.2",
    "xmltodict>=0.14.2",
]

[project.scripts]
parami = "parami.main:main"

[build-system]
requires = ["uv_build>=0.8.8,<0.9.0"]
build-backend = "uv_build"
